last

+++ init/db_dwh/create_tables.sql
@@ -21,7 +21,7 @@ create table if not exists public.curr_rates
         datetime timestamp without time zone,
         curr_id_from int references public.currencies(id),
         curr_id_to int references public.currencies(id),
-        val float
+        val numeric(15,7)
 );
 
 create unique index if not exists in_curr_rates_id on public.curr_rates using btree
